+++ pkg/domain/cron-jobs/handler.go
@@ -6,17 +6,21 @@ import (
 	"se-school-case/pkg/util/constants"
 )
 
-type CronJobsController struct {
+type CronJobsInterface interface {
+	StartScheduler()
+}
+
+type CronJobsHandler struct {
 	mailService MailInterface
 }
 
-func NewController(router *gin.Engine, mailService MailInterface) *CronJobsController {
-	ctrl := &CronJobsController{mailService: mailService}
+func NewHandler(router *gin.Engine, mailService MailInterface) *CronJobsHandler {
+	ctrl := &CronJobsHandler{mailService: mailService}
 	router.POST("/api/notify", ctrl.PostExplicitlyNotify)
 	return ctrl
 }
 
-func (c *CronJobsController) PostExplicitlyNotify(context *gin.Context) {
+func (c *CronJobsHandler) PostExplicitlyNotify(context *gin.Context) {
 	c.mailService.SendEmailToAll("Exchange rate notification", constants.TEMPLATE_PATH)
 	context.JSON(http.StatusOK, gin.H{"message": "Successfully notified all users."})
 }
